refactor(cli): split Spica adapter implementation
